bogdan/text/requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
google-api-core==1.10.0
google-auth==1.6.3
google-cloud-automl==0.2.0
googleapis-common-protos==1.5.10
grpcio==1.20.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
redis==3.2.1
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.3
Werkzeug==0.15.2
